Use explicit sbt version suffixes (013 and 100)
This slightly unifies the code visually and makes it less asymethric
