+++ test/src/test/scala/jetbrains/buildServer/sbtlogger/SbtLoggerOutputTest100.scala
@@ -3,18 +3,18 @@ package jetbrains.buildServer.sbtlogger
 import junit.framework.Assert
 import org.junit.Test
 
-import java.io.{BufferedReader, File, FileReader, IOException}
+import java.io.File
 
-class SbtLoggerOutputTest {
-
-  @Test
-  def testPluginStatus(): Unit = {
-    SbtProcess.runAndTest("sbt-teamcity-logger", testPath("compileerror"), "plugin_status_output.txt")
-  }
+class SbtLoggerOutputTest100 {
 
   @Test
   def testCompileErrorOutput(): Unit = {
     SbtProcess.runAndTest("compile", testPath("compileerror"))
+  }
+
+  @Test
+  def testPluginStatus(): Unit = {
+    SbtProcess.runAndTest("sbt-teamcity-logger", testPath("compileerror"), "plugin_status_output.txt")
   }
 
   @Test
@@ -33,23 +33,27 @@ class SbtLoggerOutputTest {
   }
 
   @Test
+  def testJUnit(): Unit = {
+    val exitCode = SbtProcess.runAndTest("test", testPath("testsupport/junit"), "output.txt")
+    // if need exit code equals 0, otherwise in TeamCity additional non-informative build problem message will appear
+    Assert.assertEquals(0, exitCode)
+  }
+
+  /*
+  @Test
   def testScalaTest(): Unit = {
     val exitCode = SbtProcess.runAndTest("test", testPath("testsupport/scalatest"), "output.txt", "output1.txt")
     // if need exit code equals 0, otherwise in TeamCity additional non-informative build problem message will appear
     Assert.assertEquals(0, exitCode)
   }
-
-  @Test
-  def testNoSbtFileInProject(): Unit = {
-    SbtProcess.runAndTest("compile", testPath("nosbtfile"))
-  }
+   */
 
   @Test
   def testWarningInspectionsInCompile(): Unit = {
     SbtProcess.runAndTestWithAdditionalParams("clean compile", "", testPath("compileInspections"))
   }
 
-  @Test
+  // todo[shkate]
   def testWarningInTestOutput(): Unit = {
     SbtProcess.runAndTest("test", testPath("TW35693"))
   }
@@ -77,21 +81,16 @@ class SbtLoggerOutputTest {
   }
 
   @Test
-  def testOtherSbtVersions(): Unit = {
-    SbtProcess.runAndTestWithAdditionalParams("sbtVersion", "--info", testPath("otherVersions"))
-  }
-
-  @Test
   def testProjectWithJavaSources(): Unit = {
     SbtProcess.runAndTestWithAdditionalParams("clean compile run", "--debug", testPath("withJavaSources"), "output.txt")
   }
 
-  @Test
+  // todo[shkate]
   def testIgnoredTest(): Unit = {
     SbtProcess.runAndTestWithAdditionalParams("--info", "test", testPath("ignoredTest"))
   }
 
-  @Test
+  // todo[shkate]
   def testNestedSuites(): Unit = {
     SbtProcess.runAndTestWithAdditionalParams("--info", "test", testPath("testsupport/nested"))
   }
@@ -101,12 +100,12 @@ class SbtLoggerOutputTest {
     SbtProcess.runAndTest("testOnly", testPath("testsupport/scalatest_TW46964"), "output.txt")
   }
 
-  @Test
+  // todo[shkate]
   def testSpec2(): Unit = {
     SbtProcess.runAndTest("testOnly", testPath("testsupport/spec2"), "output.txt")
   }
 
-  @Test
+  // todo[shkate]
   def testTW50753_initErrorInTests(): Unit = {
     SbtProcess.runAndTest("clean compile test", testPath("TW-50753_initErrorInTests"), "output.txt")
   }
@@ -132,24 +131,8 @@ class SbtLoggerOutputTest {
     )
   }
 
-  /**
-   * Service method. Allows quickly investigate test cases failed directly on TeamCity agent.
-   * Agent output should be placed in test data directory and could be checked against required output
-   */
-  @throws[IOException]
-  def testServerLogs(): Unit = {
-    val workingDir = testPath("multiproject")
-    val requiredFile = new File(workingDir + File.separator + "output.txt")
-    val serverLogs = new File(workingDir + File.separator + "server_logs.log")
-    SbtProcess.checkOutputTest(
-      new BufferedReader(new FileReader(serverLogs)),
-      new BufferedReader(new FileReader(requiredFile)),
-      null
-    )
-  }
-
   private def testPath(testRepo: String): String = {
-    val testDir = new File(new File(SbtProcess.repoRoot(), "test/testdata"), testRepo)
+    val testDir = new File(new File(SbtProcess.repoRoot(), "test/testdata/1.0"), testRepo)
     testDir.getAbsolutePath
   }
 }
